+++ src/cljs/wombats_web_client/components/modals/create_game_modal.cljs
@@ -4,7 +4,10 @@
             [cljs-time.core :as time]
             [wombats-web-client.utils.forms
              :refer [submit-modal-input
-                     cancel-modal-input]]
+                     cancel-modal-input
+                     input-error!]]
+            [wombats-web-client.utils.time
+             :refer [local-time-to-utc]]
             [wombats-web-client.utils.games
              :refer [is-private?]]
             [wombats-web-client.components.text-input
@@ -18,7 +21,8 @@
                                                      not-an-integer
                                                      max-eight
                                                      incorrect-format-colon]]
-            [wombats-web-client.events.games :refer [create-game]]))
+            [wombats-web-client.events.games :refer [create-game]]
+            [wombats-web-client.utils.functions :refer [no-blanks?]]))
 
 (defonce radios ["public" "private"])
 (defonce initial-cmpnt-state
@@ -38,44 +42,11 @@
    :password nil
    :password-error nil})
 
-(defn input-error! [check cmpnt-state]
-  (let [{:keys [key-name test-fn error]} check
-        input (key-name @cmpnt-state)
-        key-string (name key-name)
-        error-key (keyword (str key-string "-error"))]
-
-    (when (test-fn input)
-      (swap! cmpnt-state assoc error-key error))))
-
-(defn no-blanks? [value-list]
-  ;; if something returned, a field is missing a value.
-  (empty? (filter nil? value-list)))
-
 (defn get-milliseconds [min-str]
   (let [list (clojure.string/split min-str #":")
         mins (js/parseInt (first list))
         secs (js/parseInt (last list))]
     (* 1000 (+ secs (* mins 60)))))
-
-(defn get-utc [start-time]
-  (let [date-split (clojure.string/split start-time #"-")
-        year-num (js/parseInt (first date-split))
-        month-num (js/parseInt (second date-split))
-        time-string (clojure.string/split (last date-split) #"T")
-        day-num (js/parseInt (first time-string))
-        time-split (clojure.string/split (last time-string) #":")
-        hour-num (js/parseInt (first time-split))
-        min-num (js/parseInt (last time-split))
-        local-time (time/local-date-time year-num
-                                         month-num
-                                         day-num
-                                         hour-num
-                                         min-num)
-        utc-time-without-format (.toUTCIsoString
-                                 (goog.date.UtcDateTime.fromTimestamp
-                                  (.getTime local-time)) true)
-        utc-split (clojure.string/split utc-time-without-format #" ")]
-    (clojure.string/join "T" utc-split)))
 
 (defn check-for-errors [cmpnt-state]
   (let [{:keys [:game-status
@@ -133,6 +104,10 @@
     (not (clojure.string/blank? password))
     true))
 
+(defn callback-success []
+  (re-frame/dispatch [:update-modal-error nil])
+  (re-frame/dispatch [:set-modal nil]))
+
 (def callback-error
   (fn [error cmpnt-state]
     (re-frame/dispatch [:update-modal-error (get-error-message error)])
@@ -169,7 +144,7 @@
     (when ready-to-submit
       (let [intermission-ms (get-milliseconds intermission-time)
             round-ms (get-milliseconds round-time)
-            utc-time (get-utc start-time)
+            utc-time (local-time-to-utc start-time)
             game-type :high-score ;; hardcoded for now
             is-private (is-private? game-status)
             password (if is-private password "")]
@@ -183,7 +158,7 @@
                       :is-private is-private
                       :game-type game-type
                       :name game-name
-                      :on-success #(re-frame/dispatch [:set-modal nil])
+                      :on-success #(callback-success)
                       :on-error #(callback-error % cmpnt-state)})))))
 
 (defn render-password [cmpnt-state]